refactor conans->conan
