+++ conan/api/subapi/local.py
@@ -1,9 +1,8 @@
 import os
 
-from conan.api.output import ConanOutput
 from conan.cli import make_abs_path
 from conan.internal.conan_app import ConanApp
-from conans.client.cache.editable import EditablePackages
+from conan.internal.api.local.editable import EditablePackages
 from conans.client.conanfile.build import run_build_method
 from conans.client.graph.graph import CONTEXT_HOST
 from conans.client.graph.profile_node_definer import initialize_conanfile_profile
