+++ conan/tools/microsoft/msbuilddeps.py
@@ -8,7 +8,7 @@ from jinja2 import Template
 
 from conan.internal import check_duplicated_generator
 from conan.errors import ConanException
-from conans.client.generators import relativize_path
+from conan.internal.api.install.generators import relativize_path
 from conans.model.dependencies import get_transitive_requires
 from conans.util.files import load, save
