+++ test/unittests/util/local_db_test.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from conans.client.store.localdb import LocalDB
+from conan.internal.api.remotes.localdb import LocalDB
 from conan.test.utils.test_files import temp_folder
